\name{squareIt}
\alias{squareIt}


\title{Calculating The Square of a Number}
\description{
This function  calculates the square of a number. } 

\usage{
squareIt(x)

}
\arguments{
\item{x}{The number whose square is to be calculated by the function.}

}

\author{ Muege Oezluetiras, \email{mugeo@student.ethz.ch}. }


\examples{
squareIt(8) results in 8*8 or 8^2 which is 64. 
}

\keyword{square}

